+++ MyApp.Backend/MyApp.Backend/Controllers/HttpRequestParams/RemoveBrainFromCommentParam.cs
@@ -2,7 +2,7 @@
 {
     public class RemoveBrainFromCommentParam
     {
-        public int commentId { get; set; }
-        public int userId { get; set; }
+        public long commentId { get; set; }
+        public long userId { get; set; }
     }
 }
